armor rating and balance tweaks

+++ code/_core/mob/abstract/observer/_observer.dm
@@ -43,7 +43,7 @@
 
 /mob/abstract/observer/Logout()
 	. = ..()
-	qdel(src)
+	if(!src.qdeleting) qdel(src)
 	return .
 
 /mob/abstract/observer/Initialize()
